feat: add create ticket feature

+++ internal/category/handler.go
@@ -32,7 +32,7 @@ func (h *handler) ListCategories(c *echo.Context) error {
 		return response.Error(c, err)
 	}
 
-	return response.WithPagination(c, http.StatusOK, categories, params.Page, params.Limit, total)
+	return response.WithPagination(c, http.StatusOK, categories, params.Page, params.Limit, int64(total))
 }
 
 func (h *handler) CreateCategory(c *echo.Context) error {
