+++ internal/division/handler.go
@@ -32,7 +32,7 @@ func (h *handler) ListDivisions(c *echo.Context) error {
 		return response.Error(c, err)
 	}
 
-	return response.WithPagination(c, http.StatusOK, divisions, params.Page, params.Limit, total)
+	return response.WithPagination(c, http.StatusOK, divisions, params.Page, params.Limit, int64(total))
 }
 
 func (h *handler) CreateDivision(c *echo.Context) error {
